+++ src/handlers/listPet.ts
@@ -3,18 +3,32 @@ import { DynamoDB } from 'aws-sdk';
 import * as log from 'lambda-log';
 import { HttpResultV2 } from '../libs';
 
-import { HttpStatusCode, PetSortKeyMetadata, TableName } from '../models';
+import { HttpStatusCode, ListPetQuerySchema, PetSortKeyMetadata, TableName } from '../models';
 
 const dynamoDb = new DynamoDB.DocumentClient();
 
 export const listPet = async (event: APIGatewayEvent): Promise<APIGatewayProxyResultV2> => {
     log.options.meta.event = event;
 
+    const { value: query, error: queryError } = ListPetQuerySchema.validate(event.queryStringParameters, { abortEarly: false });
+    if (queryError) {
+        const arrayOfMessage: string[] = queryError.details.map(element => element.message);
+        const message = JSON.stringify({ message: arrayOfMessage });
+        log.error(message);
+        return HttpResultV2(HttpStatusCode.Invalid, { message: arrayOfMessage });
+    }
+
+    const isStatusQuery = query && query.status;
+
     const params: DynamoDB.DocumentClient.ScanInput = {
         TableName: TableName.Pet,
-        ExpressionAttributeNames: { '#t': 'type' },
-        ExpressionAttributeValues: { ':t' : PetSortKeyMetadata },
-        FilterExpression: '#t = :t'
+        ExpressionAttributeNames: isStatusQuery ? 
+            { '#t': 'type', '#s': 'status' }:
+            { '#t': 'type' },
+        ExpressionAttributeValues: isStatusQuery ? 
+            { ':t' : PetSortKeyMetadata, ':s': query.status }:
+            { ':t' : PetSortKeyMetadata },
+        FilterExpression: `#t = :t${isStatusQuery ? ' AND #s = :s' : ''}`
     };
 
     return dynamoDb.scan(params).promise().then(result => {
