+++ src/models/index.ts
@@ -1,4 +1,3 @@
 export * from './enums';
 export * from './constants';
-export * from './petSchema';
-export * from './storeSchema';
+export * from './schema';
